requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-phonenumber-field==2.1.0
django-tastypie==0.14.2
idna==2.8
phonenumbers==8.10.3
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
